Buscador funcionaba raro

+++ src/main/java/searcher/CriteriaByIdAndString.kt
@@ -6,7 +6,7 @@ class CriteriaByIdAndString(val parameter : Any?) : Criteria() {
         var parcialList : MutableCollection<Searchable?> = mutableListOf()
         var listResult : MutableCollection<Searchable?> = mutableListOf()
         when(parameter){
-            is Int -> parcialList.addAll(CriteriaById(parameter).search(toSearch)
+            is Int -> parcialList.addAll(CriteriaById(parameter.toInt()).search(toSearch)
                                                 + CriteriaByString(parameter.toString()).search(toSearch))
             is String -> parcialList = CriteriaByString(parameter).search(toSearch)
         }
